Type id aanpassen

+++ src/main/java/example/micronaut/EmployeeRepository.java
@@ -4,5 +4,5 @@ import io.micronaut.data.annotation.*;
 import io.micronaut.data.repository.CrudRepository;
 
 @Repository
-interface EmployeeRepository extends CrudRepository<Employee, Long> {
+interface EmployeeRepository extends CrudRepository<Employee, Integer> {
 }
